Move ReactGoogleMaps inside components
