+++ src/App.js
@@ -4,7 +4,7 @@ import './App.css';
 // import RoomsDropdownField from './components/RoomsDropdownField';
 
 // import ListingPage from './components/ListingPage';
-import MyMapComponent from './ReactGoogleMaps';
+import MyMapComponent from '../src/components/ReactGoogleMaps';
 
 // { /* <div className="App" >
 //         <LandingPageBody />
